Content/week6/Day4/Notifications/requirements.txt

Django==4.2.4
djangorestframework==3.14.0
pyjwt=2.8.0 --> 
djangorestframework-simplejwt=5.3.0 --> pip install djangorestframework-simplejwt
